Amend presentation-by-practice queries to use PCT IDs where possible.
This means that we can do a direct query on the presentation/pct index,
rather than having to expand the list of practices and do a huge
query join by ORs. It's a lot faster.

Chemical-by-practice and section-by-practice queries are still
expanding the list. This is because these are answered by
querying matviews, and the matviews don't have PCT fields.

We should fix this.