Import practice list sizes pre April 2013.
We obtained these from the BSA by asking them:

https://apps.nhsbsa.nhs.uk/FOI/foiRequestDetail.do?bo_id=5908

Update tests to deal with data in this format, as well as the
quarterly ongoing files from the BSA.

Don't update CCG matviews for now, we need to work out how to
extrapolate backwards for CCGs which didn't exist before April 2013.
Will discuss this with Richard Croker.