Return data for all measures in API.
Extend the `measure` and `measure_by_practice` methods so they no longer
require a `measure` parameter, but will return data for all available
measures if this parameter is not supplied.

This will be used in the practice and CCG dashboards, to show all
measures for a particular organisation.

Update tests.