Create generic 'import measures' script.
This replaces the script that was previously specific to dosulepin.

Pull out measure definitions into a data structure. We will find a
better way to store definitions in the future.

This data structure includes SQL strings: we may or may not want
to improve this in the future, depending on the kinds of queries
we end up with. Let's not optimise prematurely though.

Supply a `--measure` parameter to load data for a specific measure,
or no parameter to load data for all measures defined in the file.

Update tests.