Make all measures proportions: improve import scripts.
It vastly simplifies things, without losing much, if we treat
all measures as proportions, rather than ratios. Therefore, amend
the Rosuvastatin measure to use (Rosuvastatin + Atorvastatin) as
the denominator, not just Atorvastatin. Now our import script
only needs to handle proportions, not odds.

Amend SQL queries to return an explicit `numerator` and `denominator`
field. This assumes that we'll always be able to calculate our
`numerator` and `denominator` in SQL: if that's not the case in the
future, we'll need to review this.

Note that all measures still have an `is_percentage` field. We don't
use this while importing, since all measures are proportions, but
may use it for front-end display purposes.

Beef up the tests, including tests for practices with zero numerator and
denominator prescribing, and zero prescribing of both.

I have duplicated the test results using formulae in Google Drive: this
spreadsheet is in the shared folder, called "OpenPrescribing test
data for measures".