Import ADQs from BSA data.
We will use these to calculate ADQ-based KTTs.

We have obtained this list of ADQs by FOIing BSA. Add a management
command to import the ADQ for each presentation that is available.

This command also attempts to extract the active quantity of each
presentation, and divide the ADQ by this, to calculate the percentage of
the daily dose supplied by each presentation. This is a slightly inexact
process as the active quantity is only supplied in string form.