Add measures for CCGs.
Calculate measures depending on practice<->CCG membership at the time
of running the calculation. This simplifies things vs. maintaining a
full historic record of all CCG memberships. It should not be a major
issue in any case, as practices rarely switch CCG.

Add new test data for CCGs, and ensure percentiles and cost savings
are calculated correctly. As with practices, we reproduce these
percentiles and cost savings in the Google Drive sheet.

We may need to rejig cost saving calculations to do them separately
for practices and CCGs. Leaving this until we've discussed how we
want to display cost savings in the front-end.