Fix bug preventing tabs from working consistently
Under some circumstances, clicking on a tab was resulting in the href
for the tab itself being changed. Our logic (and the log of the
bootstrap tab implementation) relied on that href not
changing. The tag implementation uses `data-target` attributes in
preference to `href` for deciding the target of the tab switching, so we
use that instead.