Reset migrations
Django best practice is to commit migrations to git, and run migrations
on production from there.  We weren't doing this, so weren't sure if/how
production was in sync with the migrations code in git.

This commit replaces all previous migrations with a single initial
migration. We should assume that production matches this state, and
therefore run the following commands there, once this code is merged:

    echo "DELETE FROM django_migrations" | python manage.py dbshell
    python manage.py migrate --fake

This commit belongs with automated deployment changes as migrations
should now be run on each deploy.