Add a "share" button to the active tab
It should be more obvious to end users that it is possible to pass
around URLs for particular analyses.

This takes the form of a small button with a modal dropdown that shows
the URL that can be copied (plus the ability to copy by clicking a
button).

Another option would have been to use "social sharing" buttons, but it's
not clear yet if this would be the main form of sharing, as opposed to
sending things around in emails.

The button is positioned in the active tab because it was difficult to
find a place we could consistently place it within the visualisation,
because the visualisation layouts vary quite a lot (e.g. between a chart
and a map).

On clicking the copy button:

* Supported browsers will see a "Copied!" tooltip
* Unsupported browsers will see advice to use keyboard shortcuts