Finish files-to-purge functionality
We delegate the logic to a method which accepts two lists of files:
files that have changed according to git, and files that have changed on
the filesystem in the static/ directory.

The latter is a simpler hopefully less error-prone alternative to
working out individual files which may have changed as a result of
compilation.