Refactor measures code and add tests.
Make this code less of a giant hairball of doom.

Create a separate file for data logic. Add tests for this file. Keep
only presentation logic in the original file.

This changes nothing visible on the front-end, but makes the code much
more maintainable, and will be a good basis for moving to React at some
point.