Refactor measures (#92)
* Show only standard practices on measures pages.

We only use standard practices to calculate our CCG-level measures, and
therefore we should only show these practices on CCG pages.

Create tests for this.

* Fix typo

* Use correct URL attribute

* Refactor measures code and add tests.

Make this code less of a giant hairball of doom.

Create a separate file for data logic. Add tests for this file. Keep
only presentation logic in the original file.

This changes nothing visible on the front-end, but makes the code much
more maintainable, and will be a good basis for moving to React at some
point.

* Add descriptive text on measures.

Supplied by Richard Croker.

* Fix typo

* Add placeholder text.

Need to improve this, and the text on the about page. This is a placeholder till I have had a chance to discuss that with Ben.