Add missing ACE description, and link to codelist.
Richard gave us a description for ACE inhibitors which I forgot to add.
Also, Ben requested that we link to the full codelist for our
ACE/ARB/PPI/statins measures, which we store as a Google doc.

We also need to turn off auto-escaping in the Django/Handlebars
templates, so that the link renders correctly.