Add custom timeouts for different kinds of graph
In particular, a map takes ages to load.

It would be better to implement this with a more accurate xpath
selector, but I'm running short of time, so leaving this for a later
refactor.

Also conditionally size images based on the URL.  Maps in particular
need to be larger or the legend overlaps the map.