Alert when there's an inverted ratio w/ list sizes
When there are small list sizes (e.g. drug dependency specialist
practices) in an analysis, the denominator is sometimes larger than the
numerator, where it would normally be expected to be smaller.

This can skew visualisations. Let a user know when this is possibly
happening, and give them the option to hide such data points.

The decision to hide (or not) is made sticky with a cookie.