Add explanatory text on Drug Tariff search page
Closes #684

I've slightly reordered the paragraphs so that the "Search" call to
action comes directly before the search input which made more sense to
me.

I've also made it so that the text doesn't appear if there's already a
search query set, the reason being that it's quite a lot of text and if
you already know what you're looking at it's probably more distracting
than helpful. But happy to take other opinions on that.