Fix (or work around) practice search performance issue
The current list search system (used for practices, chemicals and BNF
entries) is pretty crude and mobile devices in particular struggle with
the Practices page where the results list is very large.

We should probably review the whole thing at some point, but for now we
workaround this specific issue by requiring a minimum search length of
three characters for the practice page.

Closes #865