Improve support for filtering measures by tag
As discussed in #849

We now define a human-friendly name for each tag, plus optional
description text which is shown to users when filtering. (This replaces
the hardcoded references to the "lowpriority" tag in the codebase,
thereby fixing #876).

Measures now display a list of their tags which can be clicked to filter
the displayed measures by that tag.