Fix port reuse in functional tests
Django 1.10 introduces a fix for a Windows bug which has the side-effect
of preventing us from using the same port for multiple live server
tests. (See https://code.djangoproject.com/ticket/27316)

The solution is to provide Django with a range of ports and allow it to
select a free one for each test class.