Support summing measures over all CCGs/practices
This extends the `measure_by_ccg` and `measure_by_practice` endpoints to
accept a pseudo organisation ID of "_all" which returns measures
calcuated over all CCGs/practices. More precisely:

 * Numerators and denominators are summed and the ratio is
   recalculated. (Note there are some denominators like STAR-PUs where
   summing doesn't really make sense, but hey ho.)

 * Cost savings for each centile are summed, but with negative values
   ignored. Thus the aggregate cost saving at the median represents the
   saving if all CCGs worse than the median improved and all CCGs better
   than the median stayed the same, rather than saving to be had if all
   CCGs converged on the median, which is what the simple sum would give
   us.

This "pseudo ID" technique is obviously a bit of hack, but it allows us
to get the data we want without rewriting or extending large chunks of
the API or the frontend code.