Fix trigger error in test_import_dmd
In 82ecb104 we changed test_import_dmd so that it now creates some
DMDProducts before calling import_dmd.  This caused test_import_dmd to
fail with a missing trigger in the test teardown.

Before this fix, import_dmd deleted all the dmd tables and then
recreated them.  DMDProduct has a ForeignKey to some of the dmd tables,
and so deleting the tables will cause a constraint violation if any
DMDProducts exists.  (The constraint violation is raised in the test
teardown.)

Instead of deleting and recreating the dmd tables, we now create them
only if they don't already exist.

There is a further complication in that some of the dmd tables have
corresponding models (and corresponding migrations), and so will already
exist.  The desc fields on these models were CharFields with
max_length=50, but in the tables created by import_dmd these fields have
type text.