Stop including a second copy of Bootstrap on PPU pages
The way the DataTables library was being loaded ended up including a
second copy of the Bootstrap JS & CSS. Apart from being a waste of
resources, this caused the dropdowns to break on the PPU pages.

This double include also had the (perhaps desirable) side effect of
removing the custom font from these pages. However we've since removed
that font entirely so this patch should have no visual impact.