Allow dm+d data to be imported multiple times
On each import, we first clear all the dmd_* tables (except for
dmd_product).  If we didn't do that, the INSERTs in insert_row() would
fail.  This is similar to the behaviour prior to ebb46969, when we
dropped all tables before recreating them.

We need to run both process_datafiles() and create_dmd_product() inside
the same transaction, because there are foreign keys from dmd_product to
some of the other dmd_* tables.