Ensure constraint_and_index_reconstructor is actually called
constraint_and_index_reconstructor was not used correctly in
import_measures, because although it was called and yielded, the result
was never iterated over.  This PR changes it to be called and used as a
context manager, which was the original intention.

Although this code is exercised in tests, this bug was not detected,
because a test asserted that an empty list (the list of calls) was
greater than zero.  This returns True in Python 2 (where comparing
values of different types is allowed) but raises a TypeError in Python
3.  (Additionally, the wrong thing was being patched.)

This PR fixes the code and the tests to reflect the intended behaviour,
but before merging we should decide whether this is what we actually
want, as we have survived without it for a long time.