Add Sentry's JavaScript error tracking
This involves addding the Sentry error tracking library to our JS and
then passing through the account identfier which allows Sentry to
associated these errors with the correct application.

As this identifier is exposed to the world at large it needs to be the
public one, not not the private one which contains our secret key. Sentry
provide the `raven` templatetag library which generates this public ID
automatically for us.