Fix error when resetting zoom on Analyse bar charts
This was caused by the Highcharts upgrade. Specifically, v4.2.7 has this
issue and v4.2.6 doesn't. The solution (or *a* solution) seems to be to
set the x-axis categories explicity rather than relying on each data
point having a `name` property.

It's not entirely obvious how or why this works although the upgrade
contains various changes to the label display logic for category axes so
this fix is at least plausible.

Closes #1116

Highcharts changelog:
https://www.highcharts.com/documentation/changelog#highcharts-v4.2.7

Difference between versions:
https://github.com/highcharts/highcharts/compare/v4.2.6..v4.2.7