Add a crude anti-spam mechanism to feedback form
Forms which look like they might send email to arbitrary addresses are a
magnet for bots which will happily fill them in many thousands of times
before they give up. A trick that seems to work against them is to have
one field which requires a specific value to be entered. You can even
use JS to automatically fill this field and hide the control as bots of
this sort never seem to execute JS.