Tell Travis to run builds on master branch
This is an attempt to fix the issue whereby Coveralls fails to tell us
how a given PR increases or decreases coverage. (See
https://github.com/ebmdatalab/sysadmin/issues/35). We also noticed that
the badge Coveralls generates for us says that the overall coverage
level is unknown (despite it clearly reporting coverage levels for
individual PRs).

The hypothesis is that this is due to Travis not running builds on the
master branch itself, so Coveralls has no baseline to compare to.
Travis's "branches" dashboard states that the last build on master was
two years ago, which supports this.

There is a setting in Travis which runs builds on every branch, but this
is disabled on OpenPrescribing presumably for performance reasons as it
results in two sets of builds for each PR (one on the branch as it is,
and another on how the branch would look after being merged to master).

An alternative to enabling this setting is to explicitly whitelist
branches to be built, as described here:
https://docs.travis-ci.com/user/customizing-the-build/#safelisting-or-blocklisting-branches

Hopefully this won't disrupt our normal PR testing but we shall have to
see.