Add functional tests for bookmark unsubcribing
The existing tests don't actual tests the HTML forms themselves and so
would never have caught the issue with the missing CSRF token.

We also add tests for the case where a user has only a single
subscription as the form is substantially different in this case and so
was hiding an error.