Fix bug in outlier toggle
Clicking the outlier toggle triggers a call to `map.setup` which is
asynchronous and not reentrant. This means that clicking the toggle
again while the map is still updating causes an error.

This problem happens quite frequently as the update takes a while and
blocks the UI for much of the time so it looks like nothing's happening,
leading to users clicking again and then getting an error.

As the simplest possible fix this patch just introduces a lock on the
update which allows only one to run at a time.

Note that this doesn't introduce a new jQuery dependency into the map
code: the dependency was there already, it just wasn't explicit.