Add warning to patients trying to contact their practice
Point them to NHS Choices instead.

I decided against adding another warning on the practice homepage on the
basis that in order to actually contact us the user will have to go via
the Feedback page anyway (yes, they could copy/paste the email address
but I assume the people in question aren't the sort to do this) and teh
additonal warning just creates extra noise on the page for our real
users.

Closes #1138