Redirect CCG/practice homepage requests with "tags" query
When the homepages used to show a list of all measures these URLs would
show that list filtered by the selected tags. Now that we only show two
selected measures the URLs don't work as expected. We can fix the
internal links but there are external ones which we don't control and so
it makes sense to set up redirects to handle this.

Closes #1182