Use psycopg2 binary installation
This is what we're doing currently, but unless we change the package
name we'll have to compile from source once we upgrade to v2.8.

This change is being made because the binary installation can cause
problems under certain circumstances. But this doesn't seem to have
affected us so we're just going to stick to the status quo for now.

For more information see:
http://initd.org/psycopg/articles/2018/02/08/psycopg-274-released/

Closes #1183