Update requirements.txt with new transitive requirements (#1190)
Dependabot has made various changes to versions in requirements.txt, but
has not updated transitive requirements.  As such, there are
dependencies that are missing from requirements.txt.

This means that an environment set up with pip-sync is incomplete.
(pip-sync is part of pip-tools, and makes an environment match
requirements.txt exactly.)

This PR fixes that, with the updated requirements.txt generated by
running pip-compile.

This is something we'll have to be aware of as Dependabot keeps itself
busy.