Ensure stp_id is always set on MeasureValue
It was not being set on MeasureValues corresponding to CCGs.  The change
to ccg_ratios addresses that.  We then have to change stp_ratios as the
reference to stp_id was ambiguous.

The test changes assert that every MeasureValue gets checked, as
previously the MeasureValues corresponding to practices weren't actually
getting checked.  Oops.