Use explicit "aggregate" flag rather than guesswork
In older versions of the code there wasn't an explict flag to say that
the current chart represented an aggregation of CCGs (as on the All
England page) rather than a single CCG/practice. But now there is and we
should use it as the previous approach caused bugs, for example with
legend labels.

Closes #1199