Add `quantity_means_pack` field
Usually `quantity` measures something that comes *in* a pack e.g.
"number of tablets" or "number of ml of liquid". (Note this is "pack" in
the DM+D sense were a bottle and pump are both "packs"). Occasionally
though, `quantity` measures the number of packs itself. We don't
currently have a reliable way of populating this field, but we need it
in place so we can begin populating it using various heuristics.

The specific driver for doing this comes from our attempts to calculate
the tariff cost for a given quantity of prescribing. Usually we can look
up the associated DM+D product for the VMPP, divide by the `qtyval`
field and get the number of packs dispensed (which we can then multiply
by the tariff price). However, where the quantity is *already* a number
of packs, this process produces costs several orders of magnitude too
low.

This is related to the wider issue raised in #937 that we often don't
have a good grasp of what "quantity" means.