Add management command to set `quantity_means_pack` flag
At present this command does nothing but set the flag on one specfic
presentation for which we know it needs setting and clears it on
everything else. However I think it's worth getting this command in
place now so we can start evolving it to something more clever.

One suggestion we've discussed is looking in the DM+D name to see if it
contains an "N doses" string.

Probably this command should be included in the import pipeline so it
gets automatically run whenever we import new presentations.