Add direct links to measure charts
We already supported this, but there was no easy way to find out the
correct link for a given chart. This patch adds the link into the panel
heading for each chart and, as an extra usability feature, clicking the
link temporarily replaces it with a text input with the link text
already selected and copies it to the clipboard (where the browser
supports that).

Closes #1227