Automatically approve alert signups
We still send verification emails as we're using this as a form of login
mechanism (by verifying an email address you become logged in as that
user) for managing alerts.

With the verification emails disabled you can view and manage anyone's
alerts just by entering their email address, which feels like a step too
far.