Extend implicit dataset creation logic to other BQ functions
In general, we automatically create BigQuery datasets the first time we
write to them. This doesn't make much difference in production as all
the relevant datasets exist anyway, but it does in test where we use
a new dataset for each test run. I think it also makes sense as a
general principle because it minimises the assumptions the software has
to make about its environment: it now just needs a BigQuery project to
run against, rather than a BigQuery project configured in some specific
way.