Fix bug when toggling spending/items after zooming
Doing so would trigger a bug whereby Highcharts would attempt to render
parts of the x-axis which were not associated with any CCG and, when
attempting to generate labels for these parts of the x-axis, would cause
an error in our label formatting function.

It occurred to us when investigating this that toggling between items
and spending should reset the zoom anyway, because there's no benefit to
being left zoomed in to what is now a totally different set of CCGs in
the chart. This has the side effect of making the problem go away.