Support totals over CCGs with practice level data
The Analyse form allows you to select a CCG in the practice field as a
shorthand for "all the practices in this CCG". Previously we didn't
support this when calculating totals (we'd just hide the table) but this
created an annoying inconsistency in the interface.

Note that in the case where we have practice level data but no practices
or CCGs selected we will now throw an error which nothing catches.
However the interface is supposed to prohibit this configuration so if
we do get here then throwing an error (which Sentry will alert us to) is
probably the right thing.

Additionally, if we start including other organisations like STPs then
this logic will need extending. So we loudly fail on unexpected
organisation types rather than silently give wrong results.