Fix styling and JS on measure_for_one_entity pages
This broke due to an annoyingly subtle bit of behviour from the
datatables CDN: we include "bs" in the URLs to indicate that we want the
Bootstrap compatible versions of the JS and CSS, but if you include a
version number as well (e.g. "bs-3.3.7") then it includes a copy of that
version of Bootstrap in the response. It's this second copy that causes
the breakage.

Fixes #1544