Do a little bit of normalisation of scraped names
The source website sometimes changes from using normal spaces to
non-breaking ones (\xa0).  By normalising this (and other space-related
things) before saving names to the database, we avoid importing the same
concession record twice.

(Why don't we save the fully-normalised names?  The normalisation code
will change over time.)

The difference in ncso-current.html is the change of a single normal
space to a non-breaking one.  This is not obvious from `git diff`!