Disable db_timeout decorator during API tests
This decorator is excersised by its own tests and, because configuring
the timeout interacts weirdly with Django's transaction handling, it
prevents us using the standard `TestCase` class and forces us to use the
much slower `TransactionTestCase`.

This global flag is perhaps a bit nasty, but the whole `db_timeout` code
will disappear once the next MatrixStore PR is merged so I think we can
live with it.