Support building MatrixStore file without touching BigQuery
As the MatrixStore test suite expands we need to be able to construct a
MatrixStore file from some test data without involving any external
services. This involves some small refactoring of some of the
functions in `build` so that we can do this without involving mock.

`DataFactory` is refactored to remove all BigQuery related code and
leave a class which is purely concerned with constructing test data.

Two new modules, `import_test_data_fast` and `import_test_data_full`,
provide functions for taking a `DataFactory` instance and writing its
data into a SQLite file. The fast version does this entirely in memory,
whereas the full version uses BigQuery and Google Cloud Storage.

The test suite now runs two identical sets of tests (via subclasssing)
one of which uses the fast importer and the other the full.