Generate informative names for MatrixStore files
This means that the build command no longer takes a target filename but
now creates files in the `settings.MATRIXSTORE_BUILD_DIR` directory.

The files are now named according to the following format:

  matrixstore_<LATEST_MONTH_OF_DATA>_<DATETIME_OF_BUILD>_<HASH_OF_FILE>.sqlite

The name is structured so that sorting lexically gives the file with the
latest build of the latest data. It includes a hash of the file's
contents which can be used as a cache key and also for de-duplication
(so it's easy to see if rebuilding a file has resulted in any change to
the data).