Test fast & full imports produce identical files
Identical, that is, up the level of having identical SQL dumps: we're
not testing for byte-for-byte equality as the fast build process works
on an in-memory db which can't be compared directly to a file on disk.

This also involves a small change in the practice statistics import to
make it deterministic.