Add "db" module providing entry point for the rest of the app
This module should be the primary (perhaps only) place that the rest of
the system interacts with the MatrixStore. At present is provides a
singleton instance of the currently live version of the data, and a
`group_by` function which supports grouping practice levels matrices
into different types of organisation.