Stop using third-party hosted assets, where possible
Doing so leads to issues like #1645 (i.e. mysterious breakages for some
users) and given that we host everything via Cloudflare's CDN there's no
performance advantage.

We make an exception for the Mapbox assets, because we're dependent on
their tile server anyway for the maps to work.

We also don't bother vendoring those files only loaded for IE8 as the
site is pretty much unusable on IE8 anyway.