Enable caching on requests for the entire drug tariff
On 2019-05-14 someone set up a job on Zapier which requests the entire
(35MB) drug tariff every 10 minutes. We'd like Cloudflare to cache this
for us but we don't want to cache every reponse from this endpoint as it
contains NCSO concession data which gets updated regularly. As our
internal uses of this endpoint never involve requesting the entire
tariff, a pragmatic -- if hacky -- compromise is to just cache in the
case that the request doesn't specify any BNF codes.

We set the cache timeout at 8 hours, for no particulary good reason.