Add caching for expensive All England numbers
At present the All England page contains no user-specific content and so
gets cached by Cloudflare. When we add an alert signup form it will stop
being cacheable, but it takes some expensive db queries to generate and
so it will become quite slow to load.

This patch adds application-level caching for a few of these expensive
queries which can be safely cached. Data is a cached in Redis with the
connection configured via a REDIS_URL environment variable. If no such
variable is present then the application will still work, it just won't
do any caching.

The application also looks for a SOURCE_COMMIT_ID environment variable
which is uses as part of the cache keys. This has the effect that every
deployment automatically invalidates the cache. Again, if this variable
isn't present then the application falls back to doing no caching.