Add signup form for All England alerts
This just creates the OrgBookmark entry but no alert sending mechanism
is in place yet.

The replicates the pattern used in the NCSOConcessionBookmark model
whereby null values for both practice and CCG are used to indicate an
alert for All England. While not the cleanest solution it does have the
advantages both of consistency with existing code and of allowing quite
high levels of code reuse. As discussed, there's scope for a wider
refactor of how organisations are represented within the codebase, and
so I think this is acceptable for now.