Fix triple counting of All England measure numbers
Some of the All England totals are generated by aggregating CCG level data from
the MeasureValues table. It's not possible to select CCG level data by
selecting only rows where `pct_id` is not NULL because practice level
data also has the `pct_id` set. So we have to instead select where
`practice_id` is NULL.

Previously, including the extra condition that `pct_id` be not NULL was
redundant as this was true for all rows. However with the inclusion of
STP and Regional Team data this stopped being true and our query was now
returning CCG, STP and Regional Team data which resulting in a triple
counting of some of the All England numbers.