Refactor org_type selection logic on MeasureValues
This brings into one place the logic for selecting MeasureValues by the
type of organisation to which they belong. Note that this involves a
slight change in the logic for All England numbers in that we now check
that the corresponding PCT object is a CCG and is not closed. As a
matter of fact, with our current data this makes no difference to the
totals. (I suspect that this is because such PCTs are filtered out
already by the measure import process, but I'm not 100% certain of
that.)

The newly added method is exercised by the existing `by_org` method,
which is already well covered by the existing test suite.